requirements.txt
streamlit
easyocr
pandas
numpy
Pillow
opencv-python-headlessstreamlit
easyocr
pandas
numpy
Pillow
opencv-python-headless